llm-server/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.12.1
    # via
    #   httpx
    #   starlette
    #   watchfiles
certifi==2026.1.4
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   typer-slim
    #   uvicorn
diskcache==5.6.3
    # via llama-cpp-python
fastapi==0.129.0
    # via -r requirements.in
filelock==3.24.2
    # via huggingface-hub
fsspec==2026.2.0
    # via huggingface-hub
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.2.0
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via huggingface-hub
huggingface-hub==1.4.1
    # via -r requirements.in
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
jinja2==3.1.6
    # via llama-cpp-python
llama-cpp-python==0.3.16
    # via -r requirements.in
markupsafe==3.0.3
    # via jinja2
numpy==2.4.2
    # via llama-cpp-python
packaging==26.0
    # via
    #   huggingface-hub
    #   pytest
pluggy==1.6.0
    # via pytest
pydantic==2.12.5
    # via
    #   -r requirements.in
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.0
    # via -r requirements.in
pygments==2.19.2
    # via pytest
pytest==9.0.2
    # via
    #   -r requirements.in
    #   pytest-asyncio
pytest-asyncio==1.3.0
    # via -r requirements.in
python-dotenv==1.2.1
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.22
    # via -r requirements.in
pyyaml==6.0.3
    # via
    #   huggingface-hub
    #   uvicorn
requests==2.32.5
    # via -r requirements.in
shellingham==1.5.4
    # via huggingface-hub
starlette==0.52.1
    # via fastapi
tqdm==4.67.3
    # via huggingface-hub
typer-slim==0.20.0
    # via huggingface-hub
typing-extensions==4.15.0
    # via
    #   fastapi
    #   huggingface-hub
    #   llama-cpp-python
    #   pydantic
    #   pydantic-core
    #   typer-slim
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via requests
uvicorn[standard]==0.41.0
    # via -r requirements.in
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
